# openapi.yaml
openapi: 3.1.0
info:
  title: Qarax API
  description: REST API for managing virtual machines and hypervisor hosts
  license:
    name: ''
  version: 0.1.0
paths:
...
  /storage-pools/{pool_id}/hosts:
    post:
      tags:
      - storage-pools
      operationId: attach_host
      parameters:
      - name: pool_id
        in: path
        description: Storage pool unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachHostRequest'
        required: true
      responses:
        '204':
          description: Host attached to storage pool
        '404':
          description: Storage pool or host not found
        '500':
          description: Internal server error
  /storage-pools/{pool_id}/hosts/{host_id}:
    delete:
      tags:
      - storage-pools
      operationId: detach_host
      parameters:
      - name: pool_id
        in: path
        description: Storage pool unique identifier
        required: true
        schema:
          type: string
          format: uuid
      - name: host_id
        in: path
        description: Host unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Host detached from storage pool
        '404':
          description: Host not found
        '500':
          description: Internal server error
...
components:
  schemas:
...
    AttachHostRequest:
      type: object
      required:
      - host_id
      properties:
        host_id:
          type: string
          format: uuid
...